Fix encodeURI params and some clean up (#98)
* v2.7.7

* fix encodeURI issue

* version bump
