+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ebay-node-api",
-  "version": "2.8.5",
+  "version": "2.8.6",
   "description": "Ebay node api client",
   "main": "./src/index.js",
   "homepage": "https://github.com/pajaydev/ebay-node-api",
