+++ src/finding.js
@@ -14,10 +14,8 @@ const findItemsByKeywords = function (options) {
         throw new Error('INVALID_REQUEST_PARMS --> Keyword is missing, Keyword is required');
     }
     this.options.operationName = FIND_ITEMS_BY_KEYWORD;
-    this.options.param = 'keywords';
     // support only keyword string.
     if (!options.keywords) options = { keywords: options };
-    options.keywords = encodeURIComponent(options.keywords);
     this.options.additionalParam = utils.constructAdditionalParams(options);
     const url = urlObject.buildSearchUrl(this.options);
     return getRequest(url).then((data) => {
@@ -46,9 +44,6 @@ const findItemsByCategory = function (categoryID) {
 const findCompletedItems = function (options) {
     if (!options) throw new Error('INVALID_REQUEST_PARMS --> Keyword or category ID are required.');
     if (!options.keywords && !options.categoryId) throw new Error('Keyword or category ID are required.');
-    if (options.keywords) {
-        options.keywords = encodeURIComponent(options.keywords);
-    }
     this.options.operationName = FIND_COMPLETED_ITEMS;
     this.options.additionalParam = utils.constructAdditionalParams(options);
     const url = urlObject.buildSearchUrl(this.options);
@@ -67,13 +62,9 @@ const findCompletedItems = function (options) {
  */
 const findItemsAdvanced = function (options) {
     if (!options) throw new Error('INVALID_REQUEST_PARMS --> check here for input fields https://developer.ebay.com/DevZone/finding/CallRef/findItemsAdvanced.html#Input');
-    if (options.keywords) {
-        options.keywords = encodeURIComponent(options.keywords);
-    }
     this.options.operationName = FIND_ITEMS_ADV;
     this.options.additionalParam = utils.constructAdditionalParams(options);
     const url = urlObject.buildSearchUrl(this.options);
-    console.log(url);
     return getRequest(url).then((data) => {
         return JSON.parse(data).findItemsAdvancedResponse;
     }, console.error // eslint-disable-line no-console
@@ -100,9 +91,7 @@ const findItemsByProduct = function (options) {
     let type = options.type ? options.type : 'ReferenceID';
     this.options.operationName = 'findItemsByProduct';
     this.options.additionalParam = utils.constructAdditionalParams(options);
-    let url = urlObject.buildSearchUrl(this.options);
-    url = `${url}&productId.@type=${type}`;
-    console.log(url);
+    let url = `${urlObject.buildSearchUrl(this.options)}&productId.@type=${type}`;
     return getRequest(url).then((data) => {
         return JSON.parse(data).findItemsByProductResponse;
 
@@ -115,7 +104,6 @@ const findItemsIneBayStores = function (options) {
     if (!options.storeName) throw new Error('INVALID_REQUEST_PARMS --> Store name is required.');
     this.options.operationName = FIND_EBAY_STORES;
     this.options.additionalParam = utils.constructAdditionalParams(options);
-    console.log(urlObject.buildSearchUrl(this.options));
     return getRequest(urlObject.buildSearchUrl(this.options)).then((data) => {
         return JSON.parse(data).findItemsIneBayStoresResponse;
